Added tmp tests for MainActivity and code cleanup
- Commentary part in fragments were erased
- Correction for tests in LoginActivity
