+++ app/src/main/java/com/sdp/swiftwallet/presentation/fragments/HomeFragment.java
@@ -19,10 +19,6 @@ import com.sdp.swiftwallet.WalletActivity;
 
 public class HomeFragment extends Fragment {
 
-    public HomeFragment() {
-        // Required empty public constructor
-    }
-
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
