+++ app/src/main/res/layout/fragment_profile.xml
@@ -1,5 +1,6 @@
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/profile_fragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context="com.sdp.swiftwallet.presentation.fragments.ProfileFragment">
